app/requirements.txt
apache-airflow
apache-airflow-providers-postgres
pandas
xlrd
openpyxl
psycopg2-binary
Flask
flask-cors